refactor(themes): move __tests__ out of scss dir to prevent copy
